inline DetaljertResultatFelles i utlederen
Etter at kontrollårsak-filtreringen flyttet ut var utlederen eneste bruker.
